+++ graal-js/src/com.oracle.truffle.js/src/com/oracle/truffle/js/builtins/wasm/WebAssemblyTablePrototypeBuiltins.java
@@ -52,19 +52,17 @@ import com.oracle.truffle.js.builtins.wasm.WebAssemblyTablePrototypeBuiltinsFact
 import com.oracle.truffle.js.builtins.wasm.WebAssemblyTablePrototypeBuiltinsFactory.WebAssemblyTableSetNodeGen;
 import com.oracle.truffle.js.nodes.function.JSBuiltin;
 import com.oracle.truffle.js.nodes.function.JSBuiltinNode;
+import com.oracle.truffle.js.nodes.wasm.ToJSValueNode;
 import com.oracle.truffle.js.nodes.wasm.ToWebAssemblyIndexOrSizeNode;
+import com.oracle.truffle.js.nodes.wasm.ToWebAssemblyValueNode;
 import com.oracle.truffle.js.runtime.Errors;
 import com.oracle.truffle.js.runtime.JSConfig;
 import com.oracle.truffle.js.runtime.JSContext;
 import com.oracle.truffle.js.runtime.JSRealm;
-import com.oracle.truffle.js.runtime.Strings;
 import com.oracle.truffle.js.runtime.builtins.BuiltinEnum;
-import com.oracle.truffle.js.runtime.builtins.wasm.JSWebAssembly;
-import com.oracle.truffle.js.runtime.builtins.wasm.JSWebAssemblyInstance;
 import com.oracle.truffle.js.runtime.builtins.wasm.JSWebAssemblyTable;
 import com.oracle.truffle.js.runtime.builtins.wasm.JSWebAssemblyTableObject;
-import com.oracle.truffle.js.runtime.objects.JSDynamicObject;
-import com.oracle.truffle.js.runtime.objects.Null;
+import com.oracle.truffle.js.runtime.builtins.wasm.JSWebAssemblyValueTypes;
 import com.oracle.truffle.js.runtime.objects.Undefined;
 
 public class WebAssemblyTablePrototypeBuiltins extends JSBuiltinsContainer.SwitchEnum<WebAssemblyTablePrototypeBuiltins.WebAssemblyTablePrototype> {
@@ -78,7 +76,7 @@ public class WebAssemblyTablePrototypeBuiltins extends JSBuiltinsContainer.Switc
     public enum WebAssemblyTablePrototype implements BuiltinEnum<WebAssemblyTablePrototype> {
         grow(1),
         get(1),
-        set(2);
+        set(1);
 
         private final int length;
 
@@ -102,36 +100,48 @@ public class WebAssemblyTablePrototypeBuiltins extends JSBuiltinsContainer.Switc
     protected Object createNode(JSContext context, JSBuiltin builtin, boolean construct, boolean newTarget, WebAssemblyTablePrototype builtinEnum) {
         switch (builtinEnum) {
             case grow:
-                return WebAssemblyTableGrowNodeGen.create(context, builtin, args().withThis().fixedArgs(1).createArgumentNodes(context));
+                return WebAssemblyTableGrowNodeGen.create(context, builtin, args().withThis().fixedArgs(1).varArgs().createArgumentNodes(context));
             case get:
                 return WebAssemblyTableGetNodeGen.create(context, builtin, args().withThis().fixedArgs(1).createArgumentNodes(context));
             case set:
-                return WebAssemblyTableSetNodeGen.create(context, builtin, args().withThis().fixedArgs(2).createArgumentNodes(context));
+                return WebAssemblyTableSetNodeGen.create(context, builtin, args().withThis().fixedArgs(1).varArgs().createArgumentNodes(context));
         }
         return null;
     }
 
     public abstract static class WebAssemblyTableGrowNode extends JSBuiltinNode {
         @Child ToWebAssemblyIndexOrSizeNode toDeltaNode;
+        @Child ToWebAssemblyValueNode toWebAssemblyValueNode;
         private final BranchProfile errorBranch = BranchProfile.create();
         @Child InteropLibrary tableGrowLib = InteropLibrary.getFactory().createDispatched(JSConfig.InteropLibraryLimit);
 
         public WebAssemblyTableGrowNode(JSContext context, JSBuiltin builtin) {
             super(context, builtin);
             toDeltaNode = ToWebAssemblyIndexOrSizeNode.create("WebAssembly.Table.grow(): Argument 0");
+            toWebAssemblyValueNode = ToWebAssemblyValueNode.create();
         }
 
         @Specialization
-        protected Object grow(Object thiz, Object delta) {
+        protected Object grow(Object thiz, Object delta, Object[] args) {
             if (!JSWebAssemblyTable.isJSWebAssemblyTable(thiz)) {
                 errorBranch.enter();
                 throw Errors.createTypeError("WebAssembly.Table.grow(): Receiver is not a WebAssembly.Table");
             }
+            JSWebAssemblyTableObject table = (JSWebAssemblyTableObject) thiz;
             int deltaInt = toDeltaNode.executeInt(delta);
-            Object wasmTable = ((JSWebAssemblyTableObject) thiz).getWASMTable();
+            Object wasmTable = table.getWASMTable();
+            TruffleString elementKind = table.getElementKind();
+
+            final JSRealm realm = getRealm();
+            final Object wasmValue;
+            if (args.length == 0) {
+                wasmValue = JSWebAssemblyValueTypes.getDefaultValue(realm, elementKind);
+            } else {
+                wasmValue = toWebAssemblyValueNode.execute(args[0], elementKind);
+            }
             try {
-                Object growFn = getRealm().getWASMTableGrow();
-                return tableGrowLib.execute(growFn, wasmTable, deltaInt);
+                Object growFn = realm.getWASMTableGrow();
+                return tableGrowLib.execute(growFn, wasmTable, deltaInt, wasmValue);
             } catch (InteropException ex) {
                 throw Errors.shouldNotReachHere(ex);
             } catch (AbstractTruffleException ex) {
@@ -144,14 +154,14 @@ public class WebAssemblyTablePrototypeBuiltins extends JSBuiltinsContainer.Switc
 
     public abstract static class WebAssemblyTableGetNode extends JSBuiltinNode {
         @Child ToWebAssemblyIndexOrSizeNode toIndexNode;
+        @Child ToJSValueNode toJSValueNode;
         private final BranchProfile errorBranch = BranchProfile.create();
         @Child InteropLibrary tableGetLib = InteropLibrary.getFactory().createDispatched(JSConfig.InteropLibraryLimit);
-        @Child InteropLibrary wasmFnLib = InteropLibrary.getFactory().createDispatched(JSConfig.InteropLibraryLimit);
-        @Child InteropLibrary funcTypeLib = InteropLibrary.getFactory().createDispatched(JSConfig.InteropLibraryLimit);
 
         public WebAssemblyTableGetNode(JSContext context, JSBuiltin builtin) {
             super(context, builtin);
             toIndexNode = ToWebAssemblyIndexOrSizeNode.create("WebAssembly.Table.get(): Argument 0");
+            toJSValueNode = ToJSValueNode.create();
         }
 
         @Specialization
@@ -166,12 +176,7 @@ public class WebAssemblyTablePrototypeBuiltins extends JSBuiltinsContainer.Switc
             try {
                 Object getFn = realm.getWASMTableRead();
                 Object fn = tableGetLib.execute(getFn, wasmTable, indexInt);
-                if (!wasmFnLib.isExecutable(fn)) {
-                    return Null.instance;
-                }
-                Object funcTypeFn = realm.getWASMFuncType();
-                TruffleString funcType = Strings.fromJavaString((String) funcTypeLib.execute(funcTypeFn, fn));
-                return JSWebAssemblyInstance.exportFunction(getContext(), realm, fn, funcType);
+                return toJSValueNode.execute(fn);
             } catch (InteropException ex) {
                 throw Errors.shouldNotReachHere(ex);
             } catch (AbstractTruffleException ex) {
@@ -183,34 +188,37 @@ public class WebAssemblyTablePrototypeBuiltins extends JSBuiltinsContainer.Switc
 
     public abstract static class WebAssemblyTableSetNode extends JSBuiltinNode {
         @Child ToWebAssemblyIndexOrSizeNode toIndexNode;
+        @Child ToWebAssemblyValueNode toWebAssemblyValueNode;
         private final BranchProfile errorBranch = BranchProfile.create();
         @Child InteropLibrary tableSetLib = InteropLibrary.getFactory().createDispatched(JSConfig.InteropLibraryLimit);
 
         public WebAssemblyTableSetNode(JSContext context, JSBuiltin builtin) {
             super(context, builtin);
             toIndexNode = ToWebAssemblyIndexOrSizeNode.create("WebAssembly.Table.set(): Argument 0");
+            toWebAssemblyValueNode = ToWebAssemblyValueNode.create();
         }
 
         @Specialization
-        protected Object set(Object thiz, Object index, Object value) {
+        protected Object set(Object thiz, Object index, Object[] args) {
             if (!JSWebAssemblyTable.isJSWebAssemblyTable(thiz)) {
                 errorBranch.enter();
                 throw Errors.createTypeError("WebAssembly.Table.set(): Receiver is not a WebAssembly.Table");
             }
-            Object wasmTable = ((JSWebAssemblyTableObject) thiz).getWASMTable();
+            JSWebAssemblyTableObject table = (JSWebAssemblyTableObject) thiz;
             int indexInt = toIndexNode.executeInt(index);
-            Object wasmFunction;
-            if (value == Null.instance) {
-                wasmFunction = Null.instance;
-            } else if (JSWebAssembly.isExportedFunction(value)) {
-                wasmFunction = JSWebAssembly.getExportedFunction((JSDynamicObject) value);
+            Object wasmTable = table.getWASMTable();
+            TruffleString elementKind = table.getElementKind();
+            final JSRealm realm = getRealm();
+
+            final Object wasmValue;
+            if (args.length == 0) {
+                wasmValue = JSWebAssemblyValueTypes.getDefaultValue(realm, elementKind);
             } else {
-                errorBranch.enter();
-                throw Errors.createTypeError("WebAssembly.Table.set(): Argument 1 must be null or a WebAssembly function");
+                wasmValue = toWebAssemblyValueNode.execute(args[0], elementKind);
             }
             try {
-                Object setFn = getRealm().getWASMTableWrite();
-                tableSetLib.execute(setFn, wasmTable, indexInt, wasmFunction);
+                Object setFn = realm.getWASMTableWrite();
+                tableSetLib.execute(setFn, wasmTable, indexInt, wasmValue);
             } catch (InteropException ex) {
                 throw Errors.shouldNotReachHere(ex);
             } catch (AbstractTruffleException ex) {
@@ -221,5 +229,4 @@ public class WebAssemblyTablePrototypeBuiltins extends JSBuiltinsContainer.Switc
         }
 
     }
-
 }