+++ graal-js/src/com.oracle.truffle.js/src/com/oracle/truffle/js/nodes/wasm/ToJSValueNode.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -40,16 +40,27 @@
  */
 package com.oracle.truffle.js.nodes.wasm;
 
-import com.oracle.truffle.api.dsl.GenerateUncached;
 import com.oracle.truffle.api.dsl.Specialization;
+import com.oracle.truffle.api.interop.InteropException;
+import com.oracle.truffle.api.interop.InteropLibrary;
+import com.oracle.truffle.api.interop.UnsupportedMessageException;
+import com.oracle.truffle.api.strings.TruffleString;
 import com.oracle.truffle.js.nodes.JavaScriptBaseNode;
 import com.oracle.truffle.js.runtime.BigInt;
+import com.oracle.truffle.js.runtime.Errors;
+import com.oracle.truffle.js.runtime.JSConfig;
+import com.oracle.truffle.js.runtime.JSRealm;
+import com.oracle.truffle.js.runtime.Strings;
+import com.oracle.truffle.js.runtime.builtins.wasm.JSWebAssemblyInstance;
+import com.oracle.truffle.js.runtime.objects.Null;
 
 /**
- * Implementation of ToJSValue() operation. See https://www.w3.org/TR/wasm-js-api/#tojsvalue
+ * Implementation of ToJSValue() operation. See
+ * <a href="https://www.w3.org/TR/wasm-js-api/#tojsvalue">Wasm JS-API Spec</a>
  */
-@GenerateUncached
 public abstract class ToJSValueNode extends JavaScriptBaseNode {
+    @Child InteropLibrary isFuncLib = InteropLibrary.getFactory().createDispatched(JSConfig.InteropLibraryLimit);
+    @Child InteropLibrary funcTypeLib = InteropLibrary.getFactory().createDispatched(JSConfig.InteropLibraryLimit);
 
     protected ToJSValueNode() {
     }
@@ -58,16 +69,87 @@ public abstract class ToJSValueNode extends JavaScriptBaseNode {
         return ToJSValueNodeGen.create();
     }
 
+    public static ToJSValueNode getUncached() {
+        return ToJSValueNode.Uncached.INSTANCE;
+    }
+
     public abstract Object execute(Object value);
 
     @Specialization
     public Object convert(Object value) {
-        if (value instanceof Float) {
-            return (double) (float) value; // f32
+        if (value instanceof Integer) {
+            return value;
         } else if (value instanceof Long) {
-            return BigInt.valueOf((long) value); // i64
+            return BigInt.valueOf((long) value);
+        } else if (value instanceof Float) {
+            return (double) (float) value;
+        } else if (value instanceof Double) {
+            return value;
         } else {
-            return value; // i32 or f64
+            JSRealm realm = getRealm();
+            final Object refNull = realm.getWasmRefNull();
+            if (value == refNull) {
+                return Null.instance;
+            } else {
+                Object isFuncFn = realm.getWASMIsFunc();
+                try {
+                    if ((Boolean) isFuncLib.execute(isFuncFn, value)) {
+                        Object funcTypeFn = realm.getWASMFuncType();
+                        TruffleString funcType = asTString(funcTypeLib.execute(funcTypeFn, value));
+                        return JSWebAssemblyInstance.exportFunction(realm.getContext(), realm, value, funcType);
+                    } else {
+                        return value;
+                    }
+                } catch (InteropException ex) {
+                    throw Errors.shouldNotReachHere(ex);
+                }
+            }
         }
+    }
+
+    static class Uncached extends ToJSValueNode {
+        static final Uncached INSTANCE = new Uncached();
+
+        Uncached() {
+        }
+
+        @Override
+        public Object execute(Object value) {
+            if (value instanceof Integer) {
+                return value;
+            } else if (value instanceof Long) {
+                return BigInt.valueOf((long) value);
+            } else if (value instanceof Float) {
+                return (double) (float) value;
+            } else if (value instanceof Double) {
+                return value;
+            } else {
+                JSRealm realm = getRealm();
+                final Object refNull = realm.getWasmRefNull();
+                if (value == refNull) {
+                    return Null.instance;
+                } else {
+                    Object isFuncFn = realm.getWASMIsFunc();
+                    try {
+                        if ((Boolean) InteropLibrary.getUncached().execute(isFuncFn, value)) {
+                            Object funcTypeFn = realm.getWASMFuncType();
+                            TruffleString funcType = ToJSValueNode.asTString(InteropLibrary.getUncached().execute(funcTypeFn, value));
+                            return JSWebAssemblyInstance.exportFunction(realm.getContext(), realm, value, funcType);
+                        } else {
+                            return value;
+                        }
+                    } catch (InteropException ex) {
+                        throw Errors.shouldNotReachHere(ex);
+                    }
+                }
+            }
+        }
+    }
+
+    private static TruffleString asTString(Object string) throws UnsupportedMessageException {
+        if (string instanceof String) {
+            return Strings.fromJavaString((String) string);
+        }
+        return InteropLibrary.getUncached(string).asTruffleString(string);
     }
 }