+++ graal-js/src/com.oracle.truffle.js/src/com/oracle/truffle/js/runtime/builtins/wasm/JSWebAssemblyTableObject.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -41,18 +41,25 @@
 package com.oracle.truffle.js.runtime.builtins.wasm;
 
 import com.oracle.truffle.api.object.Shape;
+import com.oracle.truffle.api.strings.TruffleString;
 import com.oracle.truffle.js.runtime.objects.JSNonProxyObject;
 
 public final class JSWebAssemblyTableObject extends JSNonProxyObject {
     private final Object wasmTable;
 
-    protected JSWebAssemblyTableObject(Shape shape, Object wasmTable) {
+    private final TruffleString elementKind;
+
+    protected JSWebAssemblyTableObject(Shape shape, Object wasmTable, TruffleString elementKind) {
         super(shape);
         this.wasmTable = wasmTable;
+        this.elementKind = elementKind;
     }
 
     public Object getWASMTable() {
         return wasmTable;
     }
 
+    public TruffleString getElementKind() {
+        return elementKind;
+    }
 }